+++ apps/api/src/routes/project/core/delete.route.ts
@@ -52,7 +52,7 @@ export default ((fastify: FastifyInstance) => ({
   handler: async (req, reply) => {
     if (!req.user) throw AppError.Unauthorized("Not logged in");
 
-    const projectService = new ProjectService(req.user.id);
+    const projectService = await ProjectService.Instance(req.user.id);
     await projectService.deleteProject({ id: (req.params as { id: string }).id });
 
     return reply.code(204).send();