+++ apps/api/src/routes/project/core/updateName.route.ts
@@ -70,7 +70,7 @@ export default ((fastify: FastifyInstance) => ({
   handler: async (req, reply) => {
     if (!req.user) throw AppError.Unauthorized("Not logged in");
 
-    const projectService = new ProjectService(req.user.id);
+    const projectService = await ProjectService.Instance(req.user.id);
     const updatedProject = await projectService.updateNameProject(
       {
         id: (req.params as { id: string }).id,