+++ apps/api/src/routes/project/members/remove.route.ts
@@ -50,7 +50,7 @@ export default ((fastify: FastifyInstance) => ({
   handler: async (req, reply) => {
     if (!req.user) throw AppError.Unauthorized("Not logged in");
 
-    const projectService = new ProjectService(req.user.id);
+    const projectService = await ProjectService.Instance(req.user.id);
     await projectService.removeMemberToProject(
       (req.params as { memberId: string }).memberId,
       (req.params as { projectId: string }).projectId,