+++ eslint.config.mjs
@@ -31,7 +31,6 @@ export default defineConfig([
       "**/env.ts",
       "**/register-paths.js",
       "**/.vercel/**",
-      "**/*todo.md",
     ],
   },